add a .clang-format file (#9154)

+++ src/modules/PositionModule.h
@@ -67,7 +67,8 @@ private:
 #if USERPREFS_EVENT_MODE
   // In event mode we want to prevent excessive position broadcasts
   // we set the minimum interval to 5m
-  const uint32_t minimumTimeThreshold = max(300000, Default::getConfiguredOrDefaultMs(config.position.broadcast_smart_minimum_interval_secs, 30));
+  const uint32_t minimumTimeThreshold =
+      max(uint32_t(300000), Default::getConfiguredOrDefaultMs(config.position.broadcast_smart_minimum_interval_secs, 30));
 #else
   const uint32_t minimumTimeThreshold = Default::getConfiguredOrDefaultMs(config.position.broadcast_smart_minimum_interval_secs, 30);
 #endif
